+++ packages/google_maps_flutter/android/src/main/java/io/flutter/plugins/googlemaps/GoogleMapOptionsSink.java
@@ -32,6 +32,8 @@ interface GoogleMapOptionsSink {
 
   void setMyLocationButtonEnabled(boolean myLocationButtonEnabled);
 
+  void setIndoorEnabled(boolean indoorEnabled);
+
   void setInitialMarkers(Object initialMarkers);
 
   void setInitialPolygons(Object initialPolygons);