+++ packages/google_maps_flutter/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 0.5.19
 
+
+* Adds support for toggling Indoor View on or off.
+
 * Allow BitmapDescriptor scaling override
+
 
 ## 0.5.18